refactor(test-critic): split core into modules
