+++ agent_s3/tools/test_critic/__init__.py
@@ -5,6 +5,10 @@ Provides both static analysis of test quality and actual test execution
 as part of the TDD/ATDD workflow.
 """
 
+from .core import TestCritic, TestType, TestVerdict, select_adapter
+from .adapters.base import Adapter
+from . import adapters, reporter
+
 __all__ = [
     'TestCritic',
     'TestType',
