+++ agent_s3/tools/test_critic/adapters/js_jest.py
@@ -13,7 +13,7 @@ import subprocess
 from pathlib import Path
 from typing import List, Optional
 
-from ..core import Adapter
+from .base import Adapter
 
 logger = logging.getLogger(__name__)
